draft1/requirements.txt
opencv-python 
face_recognition
pyserial 

